@@ -3,7 +3,6 @@
 //
 // **License:** MIT
 
-const allocBuffer = require('./util').allocBuffer
 const errors = module.exports = Object.create(null)
 const UNKNOWN_ERROR = {name: 'UNKNOWN_ERROR', code: 0xffffffff}
 
@@ -17,7 +16,7 @@ class QuicError extends Error {
   }
 
   toBuffer () {
-    let buf = allocBuffer(4)
+    let buf = Buffer.alloc(4)
     buf.writeUInt32LE(this.code)
     return buf
   }
