@@ -7,7 +7,7 @@
 const QuicID = require('./id')
 const QuicError = require('./error')
 const Visitor = require('./util').Visitor
-const allocBuffer = require('./util').allocBuffer
+
 const BUF = Symbol('QuicFrame.buf')
 const initBuf = Symbol('QuicFrame.initBuf')
 
@@ -94,7 +94,7 @@ class PaddingFrame extends QuicFrame { // PADDING Frame
   }
 
   [initBuf] () {
-    return allocBuffer(1)
+    return Buffer.alloc(1)
   }
 
   static fromBuffer () {
@@ -111,7 +111,7 @@ class ResetStreamFrame extends QuicFrame { // RST_STREAM Frame
   }
 
   [initBuf] () {
-    let buf = allocBuffer(17)
+    let buf = Buffer.alloc(17)
     buf.writeUInt8(this.type)
     this.streamID.toBuffer().copy(buf, 1, 0, 4)
     this.offset.toFullBuffer().copy(buf, 5, 0, 8)
@@ -143,7 +143,7 @@ class ConnectionCloseFrame extends QuicFrame { // CONNECTION_CLOSE frame
   [initBuf] () {
     let reasonPhrase = this.error.message
     let reasonPhraseLen = reasonPhrase ? Buffer.byteLength(reasonPhrase) : 0
-    let buf = allocBuffer(7 + reasonPhraseLen)
+    let buf = Buffer.alloc(7 + reasonPhraseLen)
     buf.writeUInt8(this.type)
     this.error.toBuffer().copy(buf, 1, 0, 4)
     buf.writeUInt16LE(reasonPhraseLen, 5, true)
@@ -179,7 +179,7 @@ class GoAwayFrame extends QuicFrame { // GOAWAY Frame
   [initBuf] () {
     let reasonPhrase = this.error.message
     let reasonPhraseLen = reasonPhrase ? Buffer.byteLength(reasonPhrase) : 0
-    let buf = allocBuffer(11 + reasonPhraseLen)
+    let buf = Buffer.alloc(11 + reasonPhraseLen)
     buf.writeUInt8(this.type)
     this.error.toBuffer().copy(buf, 1, 0, 4)
     this.streamID.toBuffer().copy(buf, 5, 0, 4)
@@ -216,7 +216,7 @@ class WindowUpdateFrame extends QuicFrame { // WINDOW_UPDATE Frame
   }
 
   [initBuf] () {
-    let buf = allocBuffer(13)
+    let buf = Buffer.alloc(13)
     buf.writeUInt8(this.type)
     this.streamID.toBuffer().copy(buf, 1, 0, 4)
     this.offset.toFullBuffer().copy(buf, 5, 0, 8)
@@ -243,7 +243,7 @@ class BlockedFrame extends QuicFrame { // BLOCKED Frame
   }
 
   [initBuf] () {
-    let buf = allocBuffer(5)
+    let buf = Buffer.alloc(5)
     buf.writeUInt8(this.type)
     this.streamID.toBuffer().copy(buf, 1, 0, 4)
     return buf
@@ -268,7 +268,7 @@ class StopWaitingFrame extends QuicFrame { // STOP_WAITING Frame
 
   [initBuf] () {
     let packetNumber = this.packetNumber.toBuffer()
-    let buf = allocBuffer(1 + packetNumber.length)
+    let buf = Buffer.alloc(1 + packetNumber.length)
     buf.writeUInt8(this.type)
     packetNumber.copy(buf, 1, 0, packetNumber.length)
     return buf
@@ -294,7 +294,7 @@ class PingFrame extends QuicFrame { // PING frame
   }
 
   [initBuf] () {
-    let buf = allocBuffer(1)
+    let buf = Buffer.alloc(1)
     buf.writeUInt8(this.type)
     return buf
   }
@@ -310,7 +310,7 @@ class CongestionFrame extends QuicFrame { // CONGESTION_FEEDBACK Frame
   }
 
   [initBuf] () {
-    let buf = allocBuffer(1)
+    let buf = Buffer.alloc(1)
     buf.writeUInt8(this.type)
     return buf
   }
@@ -351,7 +351,7 @@ class StreamFrame extends QuicFrame { // STREAM Frame
     let streamLen = this.streamID.byteLen
     let offsetLen = this.offset.byteLen
     let dataLen = this.data ? this.data.length : 0
-    let buf = allocBuffer(1 + streamLen + offsetLen + (dataLen ? (dataLen + 2) : 0))
+    let buf = Buffer.alloc(1 + streamLen + offsetLen + (dataLen ? (dataLen + 2) : 0))
     let v = new Visitor(0, 1)
     buf.writeUInt8(this.type)
     v.walk(streamLen)
