@@ -5,7 +5,7 @@
 
 const crypto = require('crypto')
 const inspect = require('util').inspect
-const allocBuffer = require('./util').allocBuffer
+
 const VAL = Symbol('QuicID.val')
 const BUF = Symbol('QuicID.buf')
 
@@ -78,7 +78,7 @@ class PacketNumber extends QuicID {
     else throw new Error('invalid PacketNumber buffer')
 
     if (buf.length !== len) {
-      this[BUF] = allocBuffer(len)
+      this[BUF] = Buffer.alloc(len)
       this[BUF].writeUIntLE(value, 0, len, true)
     }
     this[VAL] = value
@@ -94,7 +94,7 @@ class PacketNumber extends QuicID {
   }
 
   toFullBuffer () {
-    let buf = allocBuffer(8)
+    let buf = Buffer.alloc(8)
     this[BUF].copy(buf, 0, 0, this[BUF].length)
     return buf
   }
@@ -123,7 +123,7 @@ class PacketNumber extends QuicID {
     else if (value <= 0xffffffff) len = 4
     else if (value <= 0xffffffffffff) len = 6
     else throw new Error('invalid packet number')
-    let buf = allocBuffer(len)
+    let buf = Buffer.alloc(len)
     buf.writeUIntLE(value, 0, len, true)
     return new PacketNumber(buf)
   }
@@ -143,7 +143,7 @@ class StreamID extends QuicID {
     else throw new Error('invalid StreamID buffer')
 
     if (buf.length !== len) {
-      this[BUF] = allocBuffer(len)
+      this[BUF] = Buffer.alloc(len)
       this[BUF].writeUIntLE(value, 0, len, true)
     }
     this[VAL] = value
@@ -159,7 +159,7 @@ class StreamID extends QuicID {
   }
 
   toFullBuffer () {
-    let buf = allocBuffer(4)
+    let buf = Buffer.alloc(4)
     this[BUF].copy(buf, 0, 0, this[BUF].length)
     return buf
   }
@@ -185,7 +185,7 @@ class StreamID extends QuicID {
     else if (value <= 0xffffff) len = 3
     else if (value <= 0xffffffff) len = 4
     else throw new Error('invalid StreamID value')
-    let buf = allocBuffer(len)
+    let buf = Buffer.alloc(len)
     buf.writeUIntLE(value, 0, len, true)
     return new StreamID(buf)
   }
@@ -232,7 +232,7 @@ class SocketAddress extends QuicID {
 
     let buf = null
     if (family === 'IPv4') {
-      buf = allocBuffer(8)
+      buf = Buffer.alloc(8)
       buf.writeUInt16LE(0x02, 0, true)
       let offset = 2
       for (let val of address.split('.')) {
@@ -250,7 +250,7 @@ class SocketAddress extends QuicID {
         address = address.join(':')
       }
 
-      buf = allocBuffer(20)
+      buf = Buffer.alloc(20)
       buf.writeUInt16LE(0x0a, 0, true)
       let offset = 2
       for (let val of address.split(':')) {
@@ -280,7 +280,7 @@ class Offset extends QuicID {
     else throw new Error('invalid Offset buffer')
 
     if (buf.length !== len) {
-      this[BUF] = allocBuffer(len)
+      this[BUF] = Buffer.alloc(len)
       if (value) this[BUF].writeUIntLE(value, 0, len, true)
     }
     this[VAL] = value
@@ -296,7 +296,7 @@ class Offset extends QuicID {
   }
 
   toFullBuffer () {
-    let buf = allocBuffer(8)
+    let buf = Buffer.alloc(8)
     this[BUF].copy(buf, 0, 0, this[BUF].length)
     return buf
   }
@@ -322,7 +322,7 @@ class Offset extends QuicID {
     else if (value <= 0xffffffffff) len = 5
     else if (value <= 0xffffffffffff) len = 6
     else throw new Error('invalid Offset buffer')
-    let buf = allocBuffer(len)
+    let buf = Buffer.alloc(len)
     if (value) buf.writeUIntLE(value, 0, len, true)
     return new Offset(buf)
   }
