@@ -8,7 +8,6 @@ const QuicTag = require('./tag')
 // const QuicFrame = require('./frame')
 const QuicError = require('./error')
 const Visitor = require('./util').Visitor
-const allocBuffer = require('./util').allocBuffer
 
 const QUIC_VERSIONS = ['Q034']
 
@@ -69,7 +68,7 @@ class NegotiationPacket extends QuicPacket {
   }
 
   toBuffer () {
-    let buf = allocBuffer(9 + 4 * this.versions.length)
+    let buf = Buffer.alloc(9 + 4 * this.versions.length)
     buf.writeUInt8(this.flag, 0, true)
     this.connectionID.toBuffer().copy(buf, 1, 0, 8)
     let offset = 9
@@ -114,7 +113,7 @@ class ResetPacket extends QuicPacket {
     if (this.socketAddress) quicTag.setTag('CADR', this.socketAddress.toBuffer())
     let tagBuf = quicTag.toBuffer()
 
-    let buf = allocBuffer(9 + tagBuf.length)
+    let buf = Buffer.alloc(9 + tagBuf.length)
     buf.writeUInt8(this.flag, 0, true)
     this.connectionID.toBuffer().copy(buf, 1, 0, 8)
     tagBuf.copy(buf, 9, 0, tagBuf.length)
