@@ -4,7 +4,6 @@
 // **License:** MIT
 
 const Visitor = require('./util').Visitor
-const allocBuffer = require('./util').allocBuffer
 
 class QuicTag {
   constructor (name) {
@@ -28,7 +27,7 @@ class QuicTag {
     for (let key of this.keys) this.byteLen += 8 + this.values[key].length
 
     let v1 = new Visitor(0, 4)
-    let buf = allocBuffer(this.byteLen).fill(0)
+    let buf = Buffer.alloc(this.byteLen).fill(0)
     buf.write(this.name, v1.start, 4)
     v1.walk(4)
     buf.writeUInt32LE(this.keys.length, v1.start, true)
